refactor: gromosXX-compatible CLI and IO cleanup
- Move coordinate reader from md.rs to gromos-io (read_coordinates + CoordinateData)
- Add GENBOX, POSITIONRED, VELOCITYRED block support in coordinate reader
- Rewrite IMD parser for gromosXX positional format (comment headers + data lines)
- Add NTF force flags, COMTRANSROT, shake_tol to ImdParameters
- Refactor md.rs CLI to match gromosXX @-parameters:
  @topo @conf @input (required), @fin @trc @tre @trf @trv (output),
  @pttopo @posresspec @refpos @distrest @angrest @dihrest @gamd (input),
  @verb @print @version @develop (control)
- All simulation parameters now come from @input .imd/.in file
- Fix pre-existing f64/usize type errors in energy_binary and trajectory_binary tests
- 9 new/updated tests for coordinate and IMD parsing
- pair_lj reference still passes at ~1e-11 precision

+++ crates/gromos-io/src/trajectory_binary.rs
@@ -585,7 +585,7 @@ mod tests {
         // Create test configuration
         let mut config = Configuration::new(10, 1, 1);
         config.current_mut().pos = (0..10)
-            .map(|i| Vec3::new(i * 0.1, i * 0.2, i * 0.3))
+            .map(|i| Vec3::new(i as f64 * 0.1, i as f64 * 0.2, i as f64 * 0.3))
             .collect();
         config.current_mut().box_config = SimBox::rectangular(3.0, 3.0, 3.0);
 
@@ -594,7 +594,7 @@ mod tests {
             let mut writer = DcdWriter::new(temp_file, "Test trajectory").unwrap();
             for step in 0..5 {
                 writer
-                    .write_frame(step, step * 0.002, &config)
+                    .write_frame(step, step as f64 * 0.002, &config)
                     .unwrap();
             }
             writer.finish().unwrap();
